AC-999 cPanel. Parse and start yaml app

+++ AppCloud-plugin/modules/servers/KuberDock/classes/KuberDock_Addon_Product.php
@@ -90,11 +90,14 @@ class KuberDock_Addon_Product extends CL_Model {
     }
 
     /**
-     *
+     * @throws Exception
      */
     public function deletePackage()
     {
         $product = KuberDock_Product::model()->loadById($this->product_id);
+        if(!$product) {
+            throw new Exception('Product not founded');
+        }
         $product->getApi()->deletePackage($this->kuber_product_id);
     }
